# src/main/resources/openapi.yml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Telemetry Ingestion API",
    "description": "API for managing weather sensors and ingesting telemetry data.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/reports/measurements": {
      "get": {
        "tags": [
          "Reporting"
        ],
        "summary": "Query and aggregate sensor measurement data.",
        "operationId": "getAggregatedMeasurements",
        "parameters": [
          {
            "name": "sensorIds",
            "in": "query",
            "description": "Comma-separated list of sensor IDs. All sensors are included if omitted.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "minItems": 1,
              "uniqueItems": true
            },
            "style": "form",
            "explode": false
          },
          {
            "name": "metricTypes",
            "in": "query",
            "description": "Comma-separated list of metrics to include in results.",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/MetricType"
              },
              "minItems": 1,
              "uniqueItems": true
            },
            "style": "form",
            "explode": false
          },
          {
            "name": "statistic",
            "in": "query",
            "description": "The type of aggregation to perform on the metrics.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/StatisticType"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "The start date of the query time range (ISO 8601 format). Must be within one month of endDate.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "The end date of the query time range (ISO 8601 format). Must be within one month of startDate. If both are omitted, the latest data is queried.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Aggregated sensor data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AggregatedMeasurementResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input provided, e.g., invalid date range or unsupported metric type.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "One or more of the requested sensors were not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "MetricType": {
        "type": "string",
        "description": "The type of the metric being measured.",
        "example": "TEMPERATURE",
        "enum": [
          "TEMPERATURE",
          "HUMIDITY",
          "WIND_SPEED",
          "WIND_DIRECTION",
          "RAINFALL",
          "ATMOSPHERIC_PRESSURE",
          "SOLAR_RADIATION",
          "SUNLIGHT_INTENSITY",
          "UV_RADIATION",
          "NOISE_LEVEL",
          "FUNK_LEVEL"
        ]
      },
...
      "AggregatedMeasurementResponse": {
        "type": "object",
        "description": "Aggregated sensor data for a specific time range.",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AggregatedSensorResult"
            }
          },
          "queryTimestamp": {
            "type": "string",
            "format": "date-time",
            "description": "The time the query was executed."
          }
        }
      },
      "AggregatedSensorResult": {
        "type": "object",
        "description": "Aggregated data for a single sensor.",
        "properties": {
          "sensorId": {
            "type": "string",
            "description": "The unique identifier of the sensor.",
            "example": "PLANEX-HQ-TEMP-HUMIDITY"
          },
          "aggregatedMetrics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AggregatedMetric"
            }
          }
        }
      },
      "AggregatedMetric": {
        "type": "object",
        "description": "The aggregated value for a specific metric type.",
        "properties": {
          "metricType": {
            "$ref": "#/components/schemas/MetricType"
          },
          "statistic": {
            "$ref": "#/components/schemas/StatisticType"
          },
          "value": {
            "type": "number",
            "format": "double",
            "description": "The calculated aggregated value.",
            "example": 24.1
          }
        }
      },
      "StatisticType": {
        "type": "string",
        "description": "The type of statistical aggregation to apply to the data.",
        "enum": [
          "min",
          "max",
          "sum",
          "average"
        ]
      }
    }
  }
}